Move finite elements into own subpackage and modules, add caching and numba
