+++ src/planestress/analysis/finite_elements/__init__.py
@@ -1,0 +1,1 @@
+"""planestress finite elements."""
