+++ todo.md
@@ -16,10 +16,6 @@
 - All round performance improvements
 - Add benchmarks
 - Add validation tests
-- Add sparse matrices
-  - What is the best way to modify values (BCs?)
-  - Add force vector sparsity
-  - Are list comprehensions the fastest?
 - Add numba, pypardiso
 
 ## Post-processor
